+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Relay is now able to ingest pre-aggregated sessions, which will make it possible to efficiently handle applications that produce thousands of sessions per second. ([#815](https://github.com/getsentry/relay/pull/815))
 - Add protocol support for WASM. ([#852](https://github.com/getsentry/relay/pull/852))
 - Add dynamic sampling for transactions. ([#835](https://github.com/getsentry/relay/pull/835))
+- Send network outage metric on healthcheck endpoint hit. ([#856](https://github.com/getsentry/relay/pull/856))
 
 **Bug Fixes**: